elbtrace updates and zebu-sim fixes (#25305)
* AXI trace files from changqi

* added new minio for build removal

* merge conflict Adding NOC bw mon, counter, txs sch credits

* review comments incorporated for elba_barco_rings and start-hal-zebu fixed for rtl dol run

* added new APIs from TOT ASIC and libcelbisa.so for elbtrace

* MPU post processing update. SDP ring_size cfg update. Dma post-processing update

* MPU/SDP/DMA trace and post-process update
